move to seed.py
