+++ pytorch_lightning/utilities/seed.py
@@ -16,7 +16,10 @@
 import logging
 import os
 import random
-from typing import Optional
+from contextlib import contextmanager
+from random import getstate as python_get_rng_state
+from random import setstate as python_set_rng_state
+from typing import Any, Dict, Generator, Optional
 
 import numpy as np
 import torch
@@ -116,3 +119,36 @@ def pl_worker_init_function(worker_id: int, rank: Optional[int] = None) -> None:
     # use 128 bits expressed as an integer
     stdlib_seed = (stdlib_ss.generate_state(2, dtype=np.uint64).astype(object) * [1 << 64, 1]).sum()
     random.seed(stdlib_seed)
+
+
+def collect_rng_states() -> Dict[str, Any]:
+    """Collect the global random state of :mod:`torch`, :mod:`numpy` and Python."""
+    return {"torch": torch.get_rng_state(), "numpy": np.random.get_state(), "python": python_get_rng_state()}
+
+
+def set_rng_states(rng_state_dict: Dict[str, Any]) -> None:
+    """Set the global random state of :mod:`torch`, :mod:`numpy` and Python in the current process."""
+    torch.set_rng_state(rng_state_dict.get("torch"))
+    np.random.set_state(rng_state_dict.get("numpy"))
+    version, state, gauss = rng_state_dict.get("python")
+    python_set_rng_state((version, tuple(state), gauss))
+
+
+@contextmanager
+def isolate_rng() -> Generator[None, None, None]:
+    """A context manager that resets the global random state on exit to what it was before entering.
+
+    It supports isolating the states for PyTorch, Numpy, and Python built-in random number generators.
+
+    Example:
+        >>> torch.manual_seed(1)  # doctest: +ELLIPSIS
+        <torch._C.Generator object at ...>
+        >>> with isolate_rng():
+        ...     [torch.rand(1) for _ in range(3)]
+        [tensor([0.7576]), tensor([0.2793]), tensor([0.4031])]
+        >>> torch.rand(1)
+        tensor([0.7576])
+    """
+    states = collect_rng_states()
+    yield
+    set_rng_states(states)
